audit: fix int overflow UB, remove CI silent failure path
- Fix signed int32 * int32 overflow (UB) in hs-extract.cpp and
  hs-extract-batch.cpp by casting to size_t before multiply.
- Remove fork-ci.yml test-hidden-states fallback that silently
  swallowed real test failures as 'model not patched'.
  Downloaded models (Gemma 2, Qwen3.5) use patched architectures.

+++ tools/hs-extract/hs-extract.cpp
@@ -280,7 +280,7 @@ int main(int argc, char ** argv) {
         }
 
         json << "      \"values\": [";
-        for (int32_t i = 0; i < n_tokens_out * n_embd; i++) {
+        for (size_t i = 0; i < (size_t)n_tokens_out * (size_t)n_embd; i++) {
             if (i > 0) json << ", ";
             if (i > 0 && (i % n_embd) == 0) {
                 json << "\n              ";
